+++ crates/lib/src/core/parser/grammar/sequence.rs
@@ -37,7 +37,7 @@ pub struct Sequence {
     pub(crate) allow_gaps: bool,
     is_optional: bool,
     pub(crate) terminators: Vec<Arc<dyn Matchable>>,
-    cache_key: u32,
+    cache_key: u64,
 }
 
 impl Sequence {
@@ -259,7 +259,7 @@ impl Matchable for Sequence {
         })
     }
 
-    fn cache_key(&self) -> u32 {
+    fn cache_key(&self) -> u64 {
         self.cache_key
     }
 
@@ -444,7 +444,7 @@ impl Matchable for Bracketed {
         Ok(MatchResult { child_matches, ..bracketed_match })
     }
 
-    fn cache_key(&self) -> u32 {
+    fn cache_key(&self) -> u64 {
         self.this.cache_key()
     }
 }
